refactor: Comprehensive codebase improvements
This commit introduces a wide range of improvements to the codebase, addressing issues related to security, code duplication, performance, and robustness.

Key changes include:

*   **Security**: Replaced the hardcoded Google Maps API key with a more secure method using a `.env` file and the `flutter_dotenv` package.
*   **Code Quality**: Refactored the duplicated UI and logic from the create and edit alarm pages into a single, reusable `AlarmaForm` widget.
*   **Performance**: Implemented a debounce mechanism in the location autocomplete feature to reduce API calls.
*   **Robustness**: Replaced manual implementations of polyline decoding and distance calculation with functions from dedicated packages.
*   **Error Handling**: Enhanced error handling for API calls to provide better user feedback.
*   **Native Code**: Analyzed the native Android and iOS code, identifying a lack of implementation on the iOS side.
