--- lib/utils/polyline_decoder.dart
@@ -1,32 +1,0 @@
-import 'package:google_maps_flutter/google_maps_flutter.dart';
-
-List<LatLng> decodePolyline(String encoded) {
-  final points = <LatLng>[];
-  int index = 0, lat = 0, lng = 0;
-
-  while (index < encoded.length) {
-    int shift = 0, result = 0, b;
-
-    do {
-      b = encoded.codeUnitAt(index++) - 63;
-      result |= (b & 0x1f) << shift;
-      shift += 5;
-    } while (b >= 0x20);
-    lat += (result & 1) != 0 ? ~(result >> 1) : (result >> 1);
-
-    shift = 0;
-    result = 0;
-    do {
-      b = encoded.codeUnitAt(index++) - 63;
-      result |= (b & 0x1f) << shift;
-      shift += 5;
-    } while (b >= 0x20);
-    lng += (result & 1) != 0 ? ~(result >> 1) : (result >> 1);
-
-    points.add(LatLng(lat / 1e5, lng / 1e5));
-  }
-
-  return points;
-}
-
-
