+++ pubspec.yaml
@@ -25,6 +25,8 @@ dependencies:
   flutter_launcher_icons: ^0.14.3
   sqflite_common_ffi: ^2.0.0+3
   speech_to_text: ^7.2.0
+  flutter_dotenv: ^5.1.0
+  flutter_polyline_points: ^2.0.0
 
 dev_dependencies:
   flutter_test:
